design #39: 바텀시트 공통화
