--- android/app/src/main/java/com/example/passedpath/feature/main/presentation/screen/MainBottomSheetLayout.kt
@@ -1,8 +1,0 @@
-package com.example.passedpath.feature.main.presentation.screen
-
-import androidx.compose.ui.unit.dp
-
-internal val BottomSheetHiddenVisibleHeight = 72.dp
-internal val BottomSheetMiddleVisibleHeight = 332.dp
-internal val BottomSheetExpandedTopInset = 136.dp
-internal val BottomSheetFloatingPadding = 16.dp
