+++ android/app/src/main/java/com/example/passedpath/feature/main/presentation/screen/MainBottomSheetValue.kt
@@ -1,0 +1,5 @@
+package com.example.passedpath.feature.main.presentation.screen
+
+import com.example.passedpath.ui.component.bottomsheet.BaseBottomSheetValue
+
+internal typealias MainBottomSheetValue = BaseBottomSheetValue
