+++ android/app/src/main/java/com/example/passedpath/ui/component/bottomsheet/BaseBottomSheetDefaults.kt
@@ -1,0 +1,10 @@
+package com.example.passedpath.ui.component.bottomsheet
+
+import androidx.compose.ui.unit.dp
+
+object BaseBottomSheetDefaults {
+    val hiddenVisibleHeight = 72.dp
+    val middleVisibleHeight = 332.dp
+    val expandedTopInset = 136.dp
+    val floatingPadding = 16.dp
+}
