+++ android/app/src/main/java/com/example/passedpath/ui/component/bottomsheet/BaseBottomSheetTabRow.kt
@@ -1,0 +1,106 @@
+package com.example.passedpath.ui.component.bottomsheet
+
+import androidx.annotation.DrawableRes
+import androidx.annotation.StringRes
+import androidx.compose.foundation.interaction.MutableInteractionSource
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.Immutable
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.example.passedpath.ui.theme.Gray100
+import com.example.passedpath.ui.theme.Gray400
+import com.example.passedpath.ui.theme.Gray900
+import com.example.passedpath.ui.theme.Green500
+
+@Immutable
+data class BaseBottomSheetTabItem<T>(
+    val key: T,
+    @StringRes val titleResId: Int,
+    @DrawableRes val iconResId: Int
+)
+
+@Composable
+fun <T> BaseBottomSheetTabRow(
+    items: List<BaseBottomSheetTabItem<T>>,
+    selectedKey: T,
+    onTabSelected: (T) -> Unit,
+    modifier: Modifier = Modifier
+) {
+    Row(
+        modifier = modifier
+            .padding(horizontal = 16.dp)
+            .fillMaxWidth()
+            .height(58.dp)
+            .clip(RoundedCornerShape(20.dp))
+            .background(Gray100)
+            .padding(6.dp),
+        horizontalArrangement = Arrangement.spacedBy(6.dp),
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        items.forEach { item ->
+            val selected = item.key == selectedKey
+            val interactionSource = remember(item.key) { MutableInteractionSource() }
+            Surface(
+                modifier = Modifier
+                    .weight(1f)
+                    .fillMaxHeight()
+                    .clip(RoundedCornerShape(18.dp))
+                    .clickable(
+                        enabled = !selected,
+                        interactionSource = interactionSource,
+                        indication = null
+                    ) {
+                        onTabSelected(item.key)
+                    },
+                shape = RoundedCornerShape(18.dp),
+                color = if (selected) Color.White else Color.Transparent,
+                shadowElevation = if (selected) 5.dp else 0.dp
+            ) {
+                Row(
+                    modifier = Modifier.fillMaxSize(),
+                    horizontalArrangement = Arrangement.Center,
+                    verticalAlignment = Alignment.CenterVertically
+                ) {
+                    Icon(
+                        painter = painterResource(id = item.iconResId),
+                        contentDescription = null,
+                        tint = if (selected) Green500 else Gray400,
+                        modifier = Modifier.size(18.dp)
+                    )
+                    Spacer(modifier = Modifier.width(6.dp))
+                    Text(
+                        text = stringResource(item.titleResId),
+                        color = if (selected) Gray900 else Gray400,
+                        fontWeight = if (selected) FontWeight.Bold else FontWeight.SemiBold,
+                        fontSize = 16.sp,
+                        style = androidx.compose.material3.MaterialTheme.typography.bodyMedium
+                    )
+                }
+            }
+        }
+    }
+}
