+++ android/app/src/main/java/com/example/passedpath/ui/component/bottomsheet/BaseBottomSheetValue.kt
@@ -1,0 +1,7 @@
+package com.example.passedpath.ui.component.bottomsheet
+
+enum class BaseBottomSheetValue {
+    HIDDEN,
+    MIDDLE,
+    EXPANDED
+}
